+++ src/UUID.php
@@ -43,6 +43,30 @@ use DateTime;
 final class UUID extends Key implements UuidInterface
 {
     /**
+     * When this namespace is specified, the name string is a fully-qualified domain name.
+     * @link http://tools.ietf.org/html/rfc4122#appendix-C
+     */
+    const NAMESPACE_DNS = '6ba7b810-9dad-11d1-80b4-00c04fd430c8';
+
+    /**
+     * When this namespace is specified, the name string is a URL.
+     * @link http://tools.ietf.org/html/rfc4122#appendix-C
+     */
+    const NAMESPACE_URL = '6ba7b811-9dad-11d1-80b4-00c04fd430c8';
+
+    /**
+     * When this namespace is specified, the name string is an ISO OID.
+     * @link http://tools.ietf.org/html/rfc4122#appendix-C
+     */
+    const NAMESPACE_OID = '6ba7b812-9dad-11d1-80b4-00c04fd430c8';
+    
+    /**
+     * When this namespace is specified, the name string is an X.500 DN in DER or a text output format.
+     * @link http://tools.ietf.org/html/rfc4122#appendix-C
+     */
+    const NAMESPACE_X500 = '6ba7b814-9dad-11d1-80b4-00c04fd430c8';
+
+    /**
      * Instantiate an UUID object.
      *
      * @since 0.1.0